+++ packages/e2e-tests/prebundle-svelte-deps/vite.config.js
@@ -5,9 +5,12 @@ import { svelte } from '@sveltejs/vite-plugin-svelte';
 export default defineConfig({
 	plugins: [svelte()],
 	optimizeDeps: {
-		// TODO this must be excluded because nested has an scss dep that prebundle can't handle!
-		// figure out how to exclude it automatically or at least tell the user about it in a more friendly way
-		exclude: ['e2e-test-dep-scss-only']
+		exclude: [
+			// TODO this must be excluded because nested has an scss dep that prebundle can't handle!
+			// figure out how to exclude it automatically or at least tell the user about it in a more friendly way
+			'e2e-test-dep-scss-only',
+			'e2e-test-dep-svelte-hybrid'
+		]
 	},
 	build: {
 		// make build faster by skipping transforms and minification